refactor: 코드 가독성과 변경 최소화를 위해 Resume API 분리
