+++ src/main/java/com/groom/orbit/infra/ai/app/AiService.java
@@ -4,7 +4,7 @@ import com.groom.orbit.goal.goal.app.dto.response.RecommendGoalListResponseDto;
 import com.groom.orbit.goal.goal.dao.entity.MemberGoal;
 import com.groom.orbit.goal.quest.app.dto.response.RecommendQuestListResponseDto;
 import com.groom.orbit.member.member.app.dto.response.GetFeedbackResponseDto;
-import com.groom.orbit.resume.app.dto.GetResumeResponseDto;
+import com.groom.orbit.resume.application.dto.GetResumeResponseDto;
 
 public interface AiService {
