+++ src/main/java/com/groom/orbit/resume/repository/ResumeRepository.java
@@ -1,15 +1,12 @@
-package com.groom.orbit.resume.dao;
+package com.groom.orbit.resume.repository;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.groom.orbit.resume.dao.entity.Resume;
+import com.groom.orbit.resume.repository.entity.Resume;
 
 public interface ResumeRepository extends JpaRepository<Resume, Long> {
-
-  Optional<Resume> findById(Long resumeId);
 
   List<Resume> findAllByMemberId(Long memberId);
 }
